projects/marceau-solutions/labs/email-assistant/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.130.0
anthropic==0.39.0
jinja2==3.1.4
python-multipart==0.0.9
reportlab==4.2.5
